+++ buildSrc/src/main/kotlin/info/metadude/kotlin/library/sessionize/Dependencies.kt
@@ -23,10 +23,12 @@ object Libs {
         const val junitJupiter = "5.12.2"
         const val junitPlatformLauncher = "1.12.2"
         const val kotlinCoroutines = "1.10.2"
+        const val mockitoKotlin = "5.4.0"
         const val moshi = "1.15.2"
         const val okhttp = "4.12.0"
         const val retrofit = "2.11.0"
         const val threetenbp = "1.7.1"
+        const val turbine = "1.2.0"
     }
 
     const val assertjCore = "org.assertj:assertj-core:${Versions.assertjCore}"
@@ -35,6 +37,7 @@ object Libs {
     const val junitPlatformLauncher = "org.junit.platform:junit-platform-launcher:${Versions.junitPlatformLauncher}"
     const val kotlinCoroutinesCore = "org.jetbrains.kotlinx:kotlinx-coroutines-core:${Versions.kotlinCoroutines}"
     const val kotlinCoroutinesTest = "org.jetbrains.kotlinx:kotlinx-coroutines-test:${Versions.kotlinCoroutines}"
+    const val mockitoKotlin = "org.mockito.kotlin:mockito-kotlin:${Versions.mockitoKotlin}"
     const val moshi = "com.squareup.moshi:moshi:${Versions.moshi}"
     const val moshiKotlinCodegen = "com.squareup.moshi:moshi-kotlin-codegen:${Versions.moshi}"
     const val okhttp = "com.squareup.okhttp3:okhttp:${Versions.okhttp}"
@@ -42,4 +45,5 @@ object Libs {
     const val retrofit = "com.squareup.retrofit2:retrofit:${Versions.retrofit}"
     const val retrofitMoshi = "com.squareup.retrofit2:converter-moshi:${Versions.retrofit}"
     const val threetenbp = "org.threeten:threetenbp:${Versions.threetenbp}"
+    const val turbine = "app.cash.turbine:turbine:${Versions.turbine}"
 }