+++ sessionize-repositories/src/main/kotlin/info/metadude/kotlin/library/sessionize/repositories/SessionizeRepository.kt
@@ -1,0 +1,13 @@
+package info.metadude.kotlin.library.sessionize.repositories
+
+import info.metadude.kotlin.library.sessionize.repositories.models.GetConferenceDaysState
+import kotlinx.coroutines.flow.Flow
+
+interface SessionizeRepository {
+
+    suspend fun getConferenceDaysState(
+        requestETag: String,
+        lastModifiedAt: String,
+    ): Flow<GetConferenceDaysState>
+
+}